Restructured tests. Added github cache to actions. Not all tests passing yet.
