+++ .github/workflows/tests.yml
@@ -22,11 +22,10 @@ jobs:
         fetch-depth: 0
 
     - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v2
+      uses: actions/setup-python@v5
       with:
         python-version: ${{ matrix.python-version }}
-
-    - name: Install dependencies
+        cache: 'pip' # caching pip dependencies
       run: |
         python -m pip install --upgrade pip
         python -m pip install .[dev]
@@ -36,6 +35,7 @@ jobs:
       with:
         distribution: temurin
         java-version: 11
+        cache: 'sbt'
 
     - name: Build scala
       run: |
