+++ tests/test_audio.py
@@ -2,10 +2,8 @@ import itertools
 import os
 
 import numpy as np
-import pytest
 import qkeras
 import tensorflow as tf
-from tensorflow.nn import softmax
 
 from chisel4ml import generate
 from chisel4ml import optimize
@@ -17,14 +15,14 @@ from chisel4ml.preprocess.lmfe_layer import LMFELayer
 SCRIPT_DIR = os.path.dirname(os.path.realpath(__file__))
 
 
-def test_fft():
+def test_fft(request):
     test_opts_dict = {
         "window": ("hamming",),
         "tone_freq": (60,),
         "amplitude": (0.8,),
         "function": (np.cos,),
-        "frame_length": (64, 128, 256, 512, 1024),
-        "num_frames": (8, 16, 32, 64),
+        "frame_length": (128, 256, 512),
+        "num_frames": (8, 32),
     }
 
     for test_case in itertools.product(*test_opts_dict.values()):
@@ -60,15 +58,18 @@ def test_fft():
         opt_model = optimize.qkeras_model(model)
         audio_preproc = generate.circuit(
             opt_model=opt_model,
-            use_verilator=True,
-            gen_waveform=True,
+            use_verilator=request.config.getoption("--use-verilator"),
+            gen_waveform=request.config.getoption("--gen-waveform"),
+            gen_timeout_sec=request.config.getoption("--generation-timeout"),
         )
         hw_res = audio_preproc(frames, sim_timeout_sec=400) / (2**12)
         sw_res = opt_model(frames.reshape(1, num_frames, frame_length))
-        # import matplotlib.pyplot as plt
-        # plt.plot(hw_res.flatten(), color='r')
-        # plt.plot(sw_res.numpy().flatten(), color='g', linestyle='dashed')
-        # plt.show()
+        if request.config.getoption("--visualize"):
+            import matplotlib.pyplot as plt
+
+            plt.plot(hw_res.flatten(), color="r")
+            plt.plot(sw_res.numpy().flatten(), color="g", linestyle="dashed")
+            plt.show()
         assert np.allclose(
             sw_res.numpy().reshape(num_frames, frame_length),
             hw_res,
@@ -78,7 +79,7 @@ def test_fft():
         audio_preproc.delete_from_server()
 
 
-def test_fft_speech_commands(audio_data):
+def test_fft_speech_commands(request, audio_data):
     _, _, test_set, _, _, _, _ = audio_data
     model = tf.keras.Sequential()
     model.add(tf.keras.layers.Input(shape=(32, 512)))
@@ -89,7 +90,10 @@ def test_fft_speech_commands(audio_data):
     opt_model = optimize.qkeras_model(model)
 
     audio_preproc = generate.circuit(
-        opt_model=opt_model, use_verilator=True, gen_waveform=True
+        opt_model=opt_model,
+        use_verilator=request.config.getoption("--use-verilator"),
+        gen_waveform=request.config.getoption("--gen-waveform"),
+        gen_timeout_sec=request.config.getoption("--generation-timeout"),
     )
     assert audio_preproc is not None
     ts_iter = test_set.as_numpy_iterator()
@@ -97,23 +101,25 @@ def test_fft_speech_commands(audio_data):
         sample, _ = next(ts_iter)
         hw_res = audio_preproc(sample) / (2**12)
         sw_res = opt_model(sample.reshape(1, 32, 512))
-        # import matplotlib.pyplot as plt
-        # plt.plot(hw_res.flatten(), color='r')
-        # plt.plot(sw_res.numpy().flatten(), color='g', linestyle='dashed')
-        # plt.show()
+        if request.config.getoption("--visualize"):
+            import matplotlib.pyplot as plt
+
+            plt.plot(hw_res.flatten(), color="r")
+            plt.plot(sw_res.numpy().flatten(), color="g", linestyle="dashed")
+            plt.show()
         assert np.allclose(sw_res.numpy().reshape(32, 512), hw_res, atol=1, rtol=0.05)
     audio_preproc.delete_from_server()
 
 
-def test_mel_engine(audio_data):
+def test_mel_engine(request, audio_data):
     test_opts_dict = {
         "window": ("hamming",),
         "tone_freq": (60,),
         "amplitude": (0.8,),
         "function": (np.cos,),
-        "frame_length": (128, 256, 512),
+        "frame_length": (128, 512),
         "num_frames": (16, 32),
-        "num_mels": (10, 13, 15, 20, 25),
+        "num_mels": (10, 20),
     }
     for test_case in itertools.product(*test_opts_dict.values()):
         tcdict = dict(zip(test_opts_dict.keys(), test_case))
@@ -158,15 +164,18 @@ def test_mel_engine(audio_data):
         opt_model = optimize.qkeras_model(model)
         audio_preproc = generate.circuit(
             opt_model=opt_model,
-            use_verilator=True,
-            gen_waveform=True,
+            use_verilator=request.config.getoption("--use-verilator"),
+            gen_waveform=request.config.getoption("--gen-waveform"),
+            gen_timeout_sec=request.config.getoption("--generation-timeout"),
         )
         hw_res = audio_preproc(frames, sim_timeout_sec=400)
         sw_res = opt_model(frames.reshape(1, num_frames, frame_length))
-        # import matplotlib.pyplot as plt
-        # plt.plot(hw_res.flatten(), color='r')
-        # plt.plot(sw_res.numpy().flatten(), color='g', linestyle='dashed')
-        # plt.show()
+        if request.config.getoption("--visualize"):
+            import matplotlib.pyplot as plt
+
+            plt.plot(hw_res.flatten(), color="r")
+            plt.plot(sw_res.numpy().flatten(), color="g", linestyle="dashed")
+            plt.show()
         assert np.allclose(
             sw_res.numpy().flatten(),
             hw_res.flatten(),
@@ -176,7 +185,7 @@ def test_mel_engine(audio_data):
         audio_preproc.delete_from_server()
 
 
-def test_lmfe_speech_commands(audio_data):
+def test_lmfe_speech_commands(request, audio_data):
     _, _, test_set, _, _, _, _ = audio_data
     fft_layer = FFTLayer(FFTConfig(fft_size=512, num_frames=32, win_fn=np.hamming(512)))
     model = tf.keras.Sequential()
@@ -192,24 +201,29 @@ def test_lmfe_speech_commands(audio_data):
     opt_model = optimize.qkeras_model(model)
 
     audio_preproc = generate.circuit(
-        opt_model=opt_model, use_verilator=True, gen_waveform=True
+        opt_model=opt_model,
+        use_verilator=request.config.getoption("--use-verilator"),
+        gen_waveform=request.config.getoption("--gen-waveform"),
+        gen_timeout_sec=request.config.getoption("--generation-timeout"),
     )
     assert audio_preproc is not None
     ts_iter = test_set.as_numpy_iterator()
-    for _ in range(20):
+    for _ in range(10):
         sample, _ = next(ts_iter)
         fft_res = np.round(fft_layer(sample.reshape(1, 32, 512)))
         hw_res = audio_preproc(fft_res.reshape(1, 32, 512))
         sw_res = opt_model(fft_res.reshape(1, 32, 512))
-        # import matplotlib.pyplot as plt
-        # plt.plot(hw_res.flatten(), color='r')
-        # plt.plot(sw_res.numpy().flatten(), color='g', linestyle='dashed')
-        # plt.show()
+        if request.config.getoption("--visualize"):
+            import matplotlib.pyplot as plt
+
+            plt.plot(hw_res.flatten(), color="r")
+            plt.plot(sw_res.numpy().flatten(), color="g", linestyle="dashed")
+            plt.show()
         assert np.allclose(sw_res.numpy().reshape(32, 20), hw_res, atol=0, rtol=0)
     audio_preproc.delete_from_server()
 
 
-def test_preproc_speech_commands(audio_data):
+def test_preproc_speech_commands(request, audio_data):
     _, _, test_set, _, _, _, _ = audio_data
     model = tf.keras.Sequential()
     model.add(tf.keras.layers.Input(shape=(32, 512)))
@@ -225,298 +239,22 @@ def test_preproc_speech_commands(audio_data):
     opt_model = optimize.qkeras_model(model)
 
     audio_preproc = generate.circuit(
-        opt_model=opt_model, use_verilator=True, gen_waveform=True
+        opt_model=opt_model,
+        use_verilator=request.config.getoption("--use-verilator"),
+        gen_waveform=request.config.getoption("--gen-waveform"),
+        gen_timeout_sec=request.config.getoption("--generation-timeout"),
     )
     assert audio_preproc is not None
     ts_iter = test_set.as_numpy_iterator()
-    for _ in range(20):
+    for _ in range(10):
         sample, _ = next(ts_iter)
         hw_res = audio_preproc(sample)
         sw_res = opt_model(sample.reshape(1, 32, 512))
-        # import matplotlib.pyplot as plt
-        # plt.plot(hw_res.flatten(), color='r')
-        # plt.plot(sw_res.numpy().flatten(), color='g', linestyle='dashed')
-        # plt.show()
+        if request.config.getoption("--visualize"):
+            import matplotlib.pyplot as plt
+
+            plt.plot(hw_res.flatten(), color="r")
+            plt.plot(sw_res.numpy().flatten(), color="g", linestyle="dashed")
+            plt.show()
         assert np.allclose(sw_res.numpy().reshape(32, 20), hw_res, atol=1, rtol=0)
     audio_preproc.delete_from_server()
-
-
-def test_audio_classifier_no_preproc_no_bias_1st_layer(
-    qnn_audio_class_no_preproc_no_bias, audio_data_preproc
-):
-    _, _, test_set, _, _, _, _ = audio_data_preproc
-    opt_model = qnn_audio_class_no_preproc_no_bias
-    circuit = generate.circuit(
-        opt_model, use_verilator=True, gen_waveform=True, num_layers=1
-    )
-    assert circuit is not None
-    ts_iter = test_set.as_numpy_iterator()
-    for _ in range(100):
-        sample, label = next(ts_iter)
-        sw_ret = opt_model.layers[2](opt_model.layers[1](sample.reshape(1, 32, 20, 1)))
-        sw_ret = np.moveaxis(sw_ret.numpy().reshape(30, 18, 1), -1, 0)
-        hw_ret = circuit.predict(sample.reshape(1, 32, 20))
-        assert np.array_equal(hw_ret, sw_ret)
-    circuit.delete_from_server()
-
-
-def test_audio_classifier_no_preproc_no_bias_1st_2nd_layer(
-    qnn_audio_class_no_preproc_no_bias, audio_data_preproc
-):
-    _, _, test_set, _, _, _, _ = audio_data_preproc
-    opt_model = qnn_audio_class_no_preproc_no_bias
-    circuit = generate.circuit(
-        opt_model, use_verilator=True, gen_waveform=True, num_layers=2
-    )
-    assert circuit is not None
-    ts_iter = test_set.as_numpy_iterator()
-    for _ in range(500):
-        sample, label = next(ts_iter)
-        sw_ret = opt_model.layers[2](opt_model.layers[1](sample.reshape(1, 32, 20, 1)))
-        sw_ret = opt_model.layers[4](opt_model.layers[3](sw_ret))
-        sw_ret = np.moveaxis(sw_ret.numpy().reshape(28, 16, 2), -1, 0)
-        hw_ret = circuit.predict(sample.reshape(1, 32, 20))
-        assert np.array_equal(hw_ret, sw_ret)
-    circuit.delete_from_server()
-
-
-def test_audio_classifier_no_preproc_no_bias_1st_2nd_3rd_layer(
-    qnn_audio_class_no_preproc_no_bias, audio_data_preproc
-):
-    _, _, test_set, _, _, _, _ = audio_data_preproc
-    opt_model = qnn_audio_class_no_preproc_no_bias
-    circuit = generate.circuit(
-        opt_model, use_verilator=True, gen_waveform=True, num_layers=3
-    )
-    assert circuit is not None
-    ts_iter = test_set.as_numpy_iterator()
-    for _ in range(100):
-        sample, label = next(ts_iter)
-        sw_ret = opt_model.layers[2](opt_model.layers[1](sample.reshape(1, 32, 20, 1)))
-        sw_ret = opt_model.layers[4](opt_model.layers[3](sw_ret))
-        sw_ret = opt_model.layers[5](sw_ret)
-        sw_ret = np.moveaxis(sw_ret.numpy().reshape(14, 8, 2), -1, 0)
-        hw_ret = circuit.predict(sample.reshape(1, 32, 20))
-        assert np.array_equal(hw_ret, sw_ret)
-    circuit.delete_from_server()
-
-
-def test_audio_classifier_no_preproc_no_bias_1st_2nd_3rd_flatten_layer(
-    qnn_audio_class_no_preproc_no_bias, audio_data_preproc
-):
-    _, _, test_set, _, _, _, _ = audio_data_preproc
-    opt_model = qnn_audio_class_no_preproc_no_bias
-    circuit = generate.circuit(
-        opt_model, use_verilator=True, gen_waveform=True, num_layers=3
-    )
-    assert circuit is not None
-    ts_iter = test_set.as_numpy_iterator()
-    for _ in range(100):
-        sample, label = next(ts_iter)
-        sw_ret = opt_model.layers[2](opt_model.layers[1](sample.reshape(1, 32, 20, 1)))
-        sw_ret = opt_model.layers[4](opt_model.layers[3](sw_ret))
-        sw_ret = opt_model.layers[5](sw_ret)
-        sw_ret = opt_model.layers[6](sw_ret)
-        sw_ret = sw_ret.numpy()
-        hw_ret = circuit.predict(sample.reshape(1, 32, 20))
-        assert np.array_equal(hw_ret.flatten(), sw_ret.flatten())
-    circuit.delete_from_server()
-
-
-def test_audio_classifier_no_preproc_no_bias_1st_2nd_3rd_4th_layer(
-    qnn_audio_class_no_preproc_no_bias, audio_data_preproc
-):
-    _, _, test_set, _, _, _, _ = audio_data_preproc
-    opt_model = qnn_audio_class_no_preproc_no_bias
-    circuit = generate.circuit(
-        opt_model, use_verilator=True, gen_waveform=True, num_layers=4
-    )
-    assert circuit is not None
-    ts_iter = test_set.as_numpy_iterator()
-    for _ in range(100):
-        sample, label = next(ts_iter)
-        sw_ret = opt_model.layers[2](opt_model.layers[1](sample.reshape(1, 32, 20, 1)))
-        sw_ret = opt_model.layers[4](opt_model.layers[3](sw_ret))
-        sw_ret = opt_model.layers[5](sw_ret)
-        sw_ret = opt_model.layers[8](opt_model.layers[7](opt_model.layers[6](sw_ret)))
-        sw_ret = sw_ret.numpy().reshape(8)
-        hw_ret = circuit.predict(sample.reshape(1, 32, 20))
-        assert np.array_equal(hw_ret, sw_ret)
-    circuit.delete_from_server()
-
-
-def test_audio_classifier_no_preproc_no_bias(
-    qnn_audio_class_no_preproc_no_bias, audio_data_preproc
-):
-    _, _, test_set, _, _, _, _ = audio_data_preproc
-    opt_model = qnn_audio_class_no_preproc_no_bias
-    circuit = generate.circuit(opt_model, use_verilator=True, gen_waveform=True)
-    assert circuit is not None
-    ts_iter = test_set.as_numpy_iterator()
-    for _ in range(100):
-        sample, label = next(ts_iter)
-        hw_ret = circuit.predict(sample.reshape(1, 32, 20))
-        sw_ret = opt_model.predict(sample.reshape(1, 32, 20, 1))
-        print(f"hw_ret: {np.argmax(hw_ret)} - sw_ret: {np.argmax(sw_ret)}")
-        assert np.array_equal(hw_ret, sw_ret.flatten())
-    circuit.delete_from_server()
-
-
-def test_audio_classifier_no_preproc_1st_layer(
-    qnn_audio_class_no_preproc, audio_data_preproc
-):
-    _, _, test_set, _, _, _, _ = audio_data_preproc
-    opt_model = qnn_audio_class_no_preproc
-    circuit = generate.circuit(
-        opt_model, use_verilator=True, gen_waveform=True, num_layers=1
-    )
-    assert circuit is not None
-    ts_iter = test_set.as_numpy_iterator()
-    for _ in range(100):
-        sample, label = next(ts_iter)
-        sw_ret = opt_model.layers[3](opt_model.layers[2](sample.reshape(1, 32, 20, 1)))
-        sw_ret = np.moveaxis(sw_ret.numpy().reshape(30, 18, 1), -1, 0)
-        hw_ret = circuit.predict(sample.reshape(1, 32, 20))
-        assert np.array_equal(hw_ret.flatten(), sw_ret.flatten())
-    circuit.delete_from_server()
-
-
-def test_audio_classifier_no_preproc_1st_2nd_layer(
-    qnn_audio_class_no_preproc, audio_data_preproc
-):
-    _, _, test_set, _, _, _, _ = audio_data_preproc
-    opt_model = qnn_audio_class_no_preproc
-    circuit = generate.circuit(
-        opt_model, use_verilator=True, gen_waveform=True, num_layers=2
-    )
-    assert circuit is not None
-    ts_iter = test_set.as_numpy_iterator()
-    for _ in range(100):
-        sample, label = next(ts_iter)
-        sw_ret = opt_model.layers[3](opt_model.layers[2](sample.reshape(1, 32, 20, 1)))
-        sw_ret = opt_model.layers[5](opt_model.layers[4](sw_ret))
-        sw_ret = np.moveaxis(sw_ret.numpy().reshape(28, 16, 2), -1, 0)
-        hw_ret = circuit.predict(sample.reshape(1, 32, 20))
-        assert np.array_equal(hw_ret, sw_ret)
-    circuit.delete_from_server()
-
-
-def test_audio_classifier_no_preproc_1st_2nd_3rd_layer(
-    qnn_audio_class_no_preproc, audio_data_preproc
-):
-    _, _, test_set, _, _, _, _ = audio_data_preproc
-    opt_model = qnn_audio_class_no_preproc
-    circuit = generate.circuit(
-        opt_model, use_verilator=True, gen_waveform=True, num_layers=3
-    )
-    assert circuit is not None
-    ts_iter = test_set.as_numpy_iterator()
-    for _ in range(100):
-        sample, label = next(ts_iter)
-        sw_ret = opt_model.layers[3](opt_model.layers[2](sample.reshape(1, 32, 20, 1)))
-        sw_ret = opt_model.layers[5](opt_model.layers[4](sw_ret))
-        sw_ret = opt_model.layers[6](sw_ret)
-        sw_ret = np.moveaxis(sw_ret.numpy().reshape(14, 8, 2), -1, 0)
-        hw_ret = circuit.predict(sample.reshape(1, 32, 20))
-        assert np.array_equal(hw_ret, sw_ret)
-    circuit.delete_from_server()
-
-
-def test_audio_classifier_no_preproc_1st_2nd_3rd_flatten_layer(
-    qnn_audio_class_no_preproc, audio_data_preproc
-):
-    _, _, test_set, _, _, _, _ = audio_data_preproc
-    opt_model = qnn_audio_class_no_preproc
-    circuit = generate.circuit(
-        opt_model, use_verilator=True, gen_waveform=True, num_layers=3
-    )
-    assert circuit is not None
-    ts_iter = test_set.as_numpy_iterator()
-    for _ in range(100):
-        sample, label = next(ts_iter)
-        sw_ret = opt_model.layers[3](opt_model.layers[2](sample.reshape(1, 32, 20, 1)))
-        sw_ret = opt_model.layers[5](opt_model.layers[4](sw_ret))
-        sw_ret = opt_model.layers[6](sw_ret)
-        sw_ret = opt_model.layers[7](sw_ret)
-        hw_ret = circuit.predict(sample.reshape(1, 32, 20))
-        assert np.array_equal(hw_ret.flatten(), sw_ret.numpy().flatten())
-    circuit.delete_from_server()
-
-
-def test_audio_classifier_no_preproc_1st_2nd_3rd_4th_layer(
-    qnn_audio_class_no_preproc, audio_data_preproc
-):
-    _, _, test_set, _, _, _, _ = audio_data_preproc
-    opt_model = qnn_audio_class_no_preproc
-    circuit = generate.circuit(
-        opt_model, use_verilator=True, gen_waveform=True, num_layers=4
-    )
-    assert circuit is not None
-    ts_iter = test_set.as_numpy_iterator()
-    for _ in range(100):
-        sample, label = next(ts_iter)
-        sw_ret = opt_model.layers[3](opt_model.layers[2](sample.reshape(1, 32, 20, 1)))
-        sw_ret = opt_model.layers[5](opt_model.layers[4](sw_ret))
-        sw_ret = opt_model.layers[6](sw_ret)  # 14x8x2
-        sw_ret = opt_model.layers[7](sw_ret)  # 224
-        sw_ret = opt_model.layers[9](opt_model.layers[8](sw_ret))
-        hw_ret = circuit.predict(sample.reshape(1, 32, 20))
-        assert np.array_equal(hw_ret.flatten(), sw_ret.numpy().flatten())
-    circuit.delete_from_server()
-
-
-def test_audio_classifier_no_preproc(qnn_audio_class_no_preproc, audio_data_preproc):
-    _, _, test_set, _, _, _, _ = audio_data_preproc
-    opt_model = qnn_audio_class_no_preproc
-    circuit = generate.circuit(opt_model, use_verilator=True, gen_waveform=True)
-    assert circuit is not None
-    ts_iter = test_set.as_numpy_iterator()
-    for _ in range(100):
-        sample, label = next(ts_iter)
-        hw_ret = circuit.predict(sample.reshape(1, 32, 20))
-        sw_ret = opt_model.predict(sample.reshape(1, 32, 20, 1))[0]
-        print(f"hw_ret: {np.argmax(hw_ret)} - sw_ret: {np.argmax(sw_ret)}")
-        assert np.array_equal(hw_ret.flatten(), sw_ret.flatten())
-    circuit.delete_from_server()
-
-
-@pytest.mark.skip(reason="takes to long")
-def test_audio_classifier_full(qnn_audio_class, audio_data):
-    _, _, test_set, _, _, _, _ = audio_data
-    opt_model = qnn_audio_class
-    circuit = generate.circuit(opt_model, use_verilator=True, gen_waveform=False)
-    assert circuit is not None
-    ts_iter = test_set.as_numpy_iterator()
-    mistake = 0
-    for _ in range(100):
-        sample, label = next(ts_iter)
-        hw_ret = circuit.predict(sample)
-        sw_ret = opt_model.predict(sample.reshape(1, 32, 512))
-        print(f"hw_ret: {np.argmax(hw_ret)} - sw_ret: {np.argmax(sw_ret)}")
-        if np.argmax(softmax(hw_ret)) != np.argmax(softmax(sw_ret)):
-            mistake = mistake + 1
-            print("MISPREDICTION!")
-        assert mistake < 5
-        print(f"Number of mispredictions is {mistake}.")
-    circuit.delete_from_server()
-
-
-@pytest.mark.skip(reason="takes to long")
-def test_audio_classifier_big_full(qnn_audio_class_big, audio_data):
-    _, _, test_set, _, _, _, _ = audio_data
-    opt_model = qnn_audio_class_big
-    circuit = generate.circuit(opt_model, use_verilator=True, gen_waveform=False)
-    assert circuit is not None
-    ts_iter = test_set.as_numpy_iterator()
-    mistake = 0
-    for _ in range(100):
-        sample, label = next(ts_iter)
-        hw_ret = circuit.predict(sample)
-        sw_ret = opt_model.predict(sample.reshape(1, 32, 512))
-        print(f"hw_ret: {np.argmax(hw_ret)} - sw_ret: {np.argmax(sw_ret)}")
-        if np.argmax(softmax(hw_ret)) != np.argmax(softmax(sw_ret)):
-            mistake = mistake + 1
-            print("MISPREDICTION!")
-        assert mistake < 5
-        print(f"Number of mispredictions is {mistake}.")
-    circuit.delete_from_server()
