+++ tox.ini
@@ -4,13 +4,12 @@ isolated_build = true
 
 [testenv]
 extras = dev
-deps = tox-gh-actions
 run_before = make protobuf
-commands = pytest tests/
+commands = python -m pytest -svv --use-verilator --durations=10
 
 [testenv:lint]
 deps = flake8
-commands = flake8 chisel4ml tests
+commands = flake8 -v chisel4ml tests
 
 [gh-actions]
 3.8: py38
